fix(playable): composite_playable strict-include + sign-conversion 정리
.clangd `MissingIncludes: Strict` + `-Wconversion` 정책 위반 정리.
- composite_playable.h: iplayable.h / cstddef 직접 include
- composite_playable.cpp: iplayable.h / scene/actor.h / cstddef / memory / utility 직접 include
- Insert 의 pos (size_t) -> iterator difference_type 암묵 변환 회피 (static_cast<std::ptrdiff_t>)

cmake 빌드는 통과 (CXXStandard.cmake 에 -Wconversion 미설정) 했으나 IDE clangd 정책 위반.

+++ src/playable/composite_playable.cpp
@@ -1,4 +1,9 @@
 #include "playable/composite_playable.h"
+#include "playable/iplayable.h"           // SJH::Playable::IPlayable (Append/Insert/Join 시그니처)
+#include "scene/actor.h"                  // SJH::Scene::Component (dynamic_cast 대상)
+#include <cstddef>                        // std::size_t, std::ptrdiff_t
+#include <memory>                         // std::unique_ptr
+#include <utility>                        // std::move
 
 namespace SJH::Playable
 {
@@ -13,7 +18,8 @@ namespace SJH::Playable
     SequencePlayable& SequencePlayable::Insert(std::size_t pos, std::unique_ptr<IPlayable> child)
     {
         if (pos > children_.size()) pos = children_.size();
-        children_.insert(children_.begin() + pos, std::move(child));
+        // size_t → ptrdiff_t 명시 캐스트 — .clangd `-Wconversion` strict 정책 회피
+        children_.insert(children_.begin() + static_cast<std::ptrdiff_t>(pos), std::move(child));
         return *this;
     }
